refactor(dashboard): move credential API functions to api/credentials SSOT
Extract createCredential, deleteCredential, sanitizeOptionalString,
githubLoginCommand, and buildCredentialCreateRequest from the UI component
(credential-settings.ts) to the API module (api/credentials.ts) where they
belong. Components should consume domain functions, not host them.
