+++ dashboard/src/components/credential-settings.test.ts
@@ -2,16 +2,18 @@
 import { describe, expect, it } from "vitest"
 import {
   buildCredentialCreateRequest,
+  githubLoginCommand,
+  sanitizeOptionalString,
+} from "../api/credentials"
+import {
   coerceCredentialType,
   credentialStateBadgeClass,
   credentialStateLabel,
   credentialTypeBadgeClass,
   credentialTypeLabel,
-  githubLoginCommand,
   isRecord,
   normalizeCredentialsResponse,
   parseCredentialState,
-  sanitizeOptionalString,
 } from "./credential-settings"
 
 describe("coerceCredentialType", () => {
