+++ dashboard/src/components/credential-settings.ts
@@ -3,10 +3,14 @@
 
 import { html } from 'htm/preact'
 import { signal } from '@preact/signals'
-import { del, post } from '../api/core'
 import {
   fetchCredentials,
+  createCredential,
+  deleteCredential,
   coerceCredentialType,
+  sanitizeOptionalString,
+  githubLoginCommand,
+  buildCredentialCreateRequest,
   type Credential,
   type CredentialCreatePayload,
   type CredentialOauthMethod,
@@ -52,17 +56,7 @@ const addDraft = signal<CredentialCreatePayload>({
   description: '',
 })
 
-// ── API ──────────────────────────────────────────────────
-
-async function createCredential(payload: CredentialCreatePayload): Promise<void> {
-  await post('/api/v1/credentials', buildCredentialCreateRequest(payload))
-}
-
-async function deleteCredential(id: string): Promise<void> {
-  await del(`/api/v1/credentials/${encodeURIComponent(id)}`)
-}
-
-// ── Helpers ──────────────────────────────────────────────
+// ── UI Helpers ────────────────────────────────────────────
 
 export function credentialTypeLabel(type: CredentialType): string {
   switch (type) {
@@ -106,41 +100,6 @@ export function credentialStateBadgeClass(state: CredentialState | null | undefi
       return 'bg-[var(--color-bg-elevated)] text-text-dim border-[var(--color-border-default)]'
     default:
       return 'bg-[var(--color-bg-elevated)] text-text-muted border-[var(--color-border-default)]'
-  }
-}
-
-export function sanitizeOptionalString(value: string | null | undefined): string | null {
-  const trimmed = value?.trim() ?? ''
-  return trimmed === '' ? null : trimmed
-}
-
-function shellQuote(value: string): string {
-  return `'${value.split("'").join("'\\''")}'`
-}
-
-export function githubLoginCommand(ghConfigDir: string | null | undefined): string | null {
-  const dir = sanitizeOptionalString(ghConfigDir)
-  if (!dir) return null
-  return `GH_CONFIG_DIR=${shellQuote(dir)} gh auth login --hostname github.com --git-protocol https --web --clipboard`
-}
-
-export function buildCredentialCreateRequest(payload: CredentialCreatePayload): Record<string, unknown> {
-  const ghConfigDir = sanitizeOptionalString(payload.gh_config_dir)
-  const sshKeyPath = sanitizeOptionalString(payload.ssh_key_path)
-  const gpgKeyId = sanitizeOptionalString(payload.gpg_key_id)
-  const oauthMethod =
-    payload.type === 'github'
-      ? payload.oauth_method === 'with_token' ? 'with_token' : 'web'
-      : 'web'
-  return {
-    id: payload.id.trim(),
-    cred_type: payload.type,
-    username: (payload.username || payload.name).trim(),
-    gh_config_dir: ghConfigDir,
-    ssh_key_path: sshKeyPath,
-    gpg_key_id: gpgKeyId,
-    oauth_method: oauthMethod,
-    token: oauthMethod === 'with_token' ? sanitizeOptionalString(payload.token) : null,
   }
 }
